# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Polly",
    "description": "configurable poll distant resources",
    "version": "1.0.0"
  },
  "host": "parallel-scarlet-juravenator.glitch.me",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/Module/insert": {
      "post": {
        "tags": [
          "Modules"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Module",
            "in": "body",
            "description": "Add a Module",
            "schema": {
              "$ref": "#/definitions/Module"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Module/updateSingle": {
      "post": {
        "tags": [
          "Modules"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Module",
            "in": "body",
            "description": "Update a Module",
            "schema": {
              "$ref": "#/definitions/Module"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Module/deleteSingle": {
      "delete": {
        "tags": [
          "Modules"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Module",
            "in": "body",
            "description": "Delete Module",
            "schema": {
              "$ref": "#/definitions/Module"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Module/getById/{module_id}": {
      "get": {
        "tags": [
          "Modules"
        ],
        "description": "get details of Module by module_id",
        "parameters": [
          {
            "name": "module_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Module successfully obtained.",
            "schema": {
              "$ref": "#/definitions/Module"
            }
          }
        }
      }
    },
...
    "/Module/getBy_owner_group_id/{owner_group_id}": {
      "get": {
        "tags": [
          "Modules"
        ],
        "description": "get all Modules by owner_group_id",
        "parameters": [
          {
            "name": "owner_group_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Module/countBy_owner_group_id/{owner_group_id}": {
      "get": {
        "tags": [
          "Modules"
        ],
        "description": "count all Modules by owner_group_id",
        "parameters": [
          {
            "name": "owner_group_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/Subscriber/insert": {
      "post": {
        "tags": [
          "Subscribers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Subscriber",
            "in": "body",
            "description": "Add a Subscriber",
            "schema": {
              "$ref": "#/definitions/Subscriber"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Subscriber/updateSingle": {
      "post": {
        "tags": [
          "Subscribers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Subscriber",
            "in": "body",
            "description": "Update a Subscriber",
            "schema": {
              "$ref": "#/definitions/Subscriber"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Subscriber/deleteSingle": {
      "delete": {
        "tags": [
          "Subscribers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Subscriber",
            "in": "body",
            "description": "Delete Subscriber",
            "schema": {
              "$ref": "#/definitions/Subscriber"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Subscriber/getById/{watcher_id}": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "get details of Subscriber by watcher_id",
        "parameters": [
          {
            "name": "watcher_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Subscriber successfully obtained.",
            "schema": {
              "$ref": "#/definitions/Subscriber"
            }
          }
        }
      }
    },
...
    "/Subscriber/getBy_subscriber_id/{subscriber_id}": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "get all Subscribers by subscriber_id",
        "parameters": [
          {
            "name": "subscriber_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Subscriber/countBy_subscriber_id/{subscriber_id}": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "count all Subscribers by subscriber_id",
        "parameters": [
          {
            "name": "subscriber_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Subscriber/getBy_owner_group_id/{owner_group_id}": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "get all Subscribers by owner_group_id",
        "parameters": [
          {
            "name": "owner_group_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Subscriber/countBy_owner_group_id/{owner_group_id}": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "count all Subscribers by owner_group_id",
        "parameters": [
          {
            "name": "owner_group_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/Watcher/insert": {
      "post": {
        "tags": [
          "Watchers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Watcher",
            "in": "body",
            "description": "Add a Watcher",
            "schema": {
              "$ref": "#/definitions/Watcher"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Watcher/updateSingle": {
      "post": {
        "tags": [
          "Watchers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Watcher",
            "in": "body",
            "description": "Update a Watcher",
            "schema": {
              "$ref": "#/definitions/Watcher"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Watcher/deleteSingle": {
      "delete": {
        "tags": [
          "Watchers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Watcher",
            "in": "body",
            "description": "Delete Watcher",
            "schema": {
              "$ref": "#/definitions/Watcher"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Watcher/getById/{watcher_id}": {
      "get": {
        "tags": [
          "Watchers"
        ],
        "description": "get details of Watcher by watcher_id",
        "parameters": [
          {
            "name": "watcher_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Watcher successfully obtained.",
            "schema": {
              "$ref": "#/definitions/Watcher"
            }
          }
        }
      }
    },
...
    "/Watcher/getBy_owner_group_id/{owner_group_id}": {
      "get": {
        "tags": [
          "Watchers"
        ],
        "description": "get all Watchers by owner_group_id",
        "parameters": [
          {
            "name": "owner_group_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Watcher/countBy_owner_group_id/{owner_group_id}": {
      "get": {
        "tags": [
          "Watchers"
        ],
        "description": "count all Watchers by owner_group_id",
        "parameters": [
          {
            "name": "owner_group_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "Watcher": {
      "type": "object",
      "properties": {
        "watcher_id": {
          "type": "string",
          "example": "autogenerated as uuid(32) by sql engine"
        },
        "name": {
          "type": "string",
          "example": ""
        },
        "owner_group_id": {
          "type": "string",
          "example": ""
        },
        "source": {
          "type": "string",
          "example": "module.exports≈{\n        main({modules,data}){\n          // returns a list of watch results\n          // the list is persisted as a watch\n          // each watch must have a unique watch_id\n          // and a payload json object\n          return [{watch_id:444,payload:444}]\n        },\n      }"
        },
        "poll_config": {
          "type": "string",
          "example": "{\n        type:\"DailyDistributionFourier\",\n        coefficients:[1,-3,5,-7,9,-11,13],\n      }"
        }
      }
    },
    "Subscriber": {
      "type": "object",
      "properties": {
        "watcher_id": {
          "type": "string",
          "example": "autogenerated as uuid(32) by sql engine"
        },
        "subscriber_id": {
          "type": "string",
          "example": ""
        },
        "owner_group_id": {
          "type": "string",
          "example": ""
        },
        "source": {
          "type": "string",
          "example": "module.exports≈{\n        main({modules,data}){\n          return [{val:444}]\n        },\n      }"
        }
      }
    },
...
    "Module": {
      "type": "object",
      "properties": {
        "module_id": {
          "type": "string",
          "example": "autogenerated as uuid(32) by sql engine"
        },
        "owner_group_id": {
          "type": "string",
          "example": ""
        },
        "npm_module": {
          "type": "string",
          "example": "moment"
        },
        "source": {
          "type": "string",
          "example": "module.exports≈{\n        /* this function signature*/\n        main:({modules,metadata})=>{\n          return [{val:444}]\n        },\n      }"
        }
      }
    },
...
  }
}